Fix another warning

+++ Config.C
@@ -192,7 +192,7 @@ void DynamicConfig::update(char *string)
 
 	if (*s != '\0') {
 	    fprintf(stderr, "\nwmx: Dynamic configuration error: "
-		    "`%s' @ position %d", s, string - s);
+		    "`%s' @ position %d", s, (int)(string - s));
 	}
 
     } while ((s = strtok(NULL, "/")));
